… VisualFSharp.sln
… tests/benchmarks/Benchmarks.sln
… tests/benchmarks/CompilerServiceBenchmarks/Program.fs
… tests/benchmarks/MicroPerf/CS/MicroPerfCSharp.cs
… tests/benchmarks/MicroPerf/CS/MicroPerfCSharp.csproj
… tests/benchmarks/MicroPerf/MicroPerf.fsproj
… tests/fsharp/Compiler/CodeGen/EmittedIL/BooleanLogic.fs
… tests/fsharp/Compiler/CodeGen/EmittedIL/ComputedListExpressions.fs
… tests/fsharp/Compiler/CodeGen/EmittedIL/Mutation.fs
… tests/fsharp/Compiler/Stress/LargeExprTests.fs
… tests/fsharp/FSharpSuite.Tests.fsproj
… tests/fsharp/core/printf/test.fsx
… tests/fsharpqa/Source/CodeGen/EmittedIL/CCtorDUWithMember/CCtorDUWithMember01.il.bsl
… tests/fsharpqa/Source/CodeGen/EmittedIL/CCtorDUWithMember/CCtorDUWithMember02.il.bsl
… tests/fsharpqa/Source/CodeGen/EmittedIL/CCtorDUWithMember/CCtorDUWithMember03.il.bsl
… tests/fsharpqa/Source/CodeGen/EmittedIL/CCtorDUWithMember/CCtorDUWithMember04.il.bsl
… tests/fsharpqa/Source/CodeGen/EmittedIL/GeneratedIterators/GenIter01.il.bsl
… tests/fsharpqa/Source/CodeGen/EmittedIL/GeneratedIterators/GenIter02.il.bsl
… tests/fsharpqa/Source/CodeGen/EmittedIL/GeneratedIterators/GenIter03.il.bsl
… tests/fsharpqa/Source/CodeGen/EmittedIL/GeneratedIterators/GenIter04.il.bsl
… tests/fsharpqa/Source/CodeGen/EmittedIL/InequalityComparison/InequalityComparison01.il.bsl
… tests/fsharpqa/Source/CodeGen/EmittedIL/InequalityComparison/InequalityComparison02.il.bsl
… tests/fsharpqa/Source/CodeGen/EmittedIL/InequalityComparison/InequalityComparison03.il.bsl
… tests/fsharpqa/Source/CodeGen/EmittedIL/InequalityComparison/InequalityComparison04.il.bsl
… tests/fsharpqa/Source/CodeGen/EmittedIL/InequalityComparison/InequalityComparison05.il.bsl
… tests/fsharpqa/Source/CodeGen/EmittedIL/ListExpressionStepping/ListExpressionSteppingTest1.il.bsl
… tests/fsharpqa/Source/CodeGen/EmittedIL/ListExpressionStepping/ListExpressionSteppingTest2.il.bsl
… tests/fsharpqa/Source/CodeGen/EmittedIL/ListExpressionStepping/ListExpressionSteppingTest3.il.bsl
… tests/fsharpqa/Source/CodeGen/EmittedIL/ListExpressionStepping/ListExpressionSteppingTest4.il.bsl
… tests/fsharpqa/Source/CodeGen/EmittedIL/ListExpressionStepping/ListExpressionSteppingTest5.il.bsl