@dsyme 2021-06-01
Note that the slowdowns reported by Bartosz in the tweet were more dramatic than those shown here. 

I'm going to work on this a little more. The codegen in the second case is still not perfect, see the corresponding C# codegen here:

https://sharplab.io/#v2:C4LglgNgPgAgTARgLACgYGYAE9MGFMDeqmJ2WYAdsJgLIAUl1AHgJSHGmlgBmmdTmALyDMCTFCiYBwzHDYdOimAHZRAbgWcAphADOWzJsUkefaSKwSpQkQBZ5KY0tVwNj4zv2H3Tkiszobk6aAL6oIUA

```
        IL_0000: ldarg.1
        IL_0001: ldc.i4.1
        IL_0002: beq.s IL_0008

        IL_0004: ldarg.1
        IL_0005: ldc.i4.2
        IL_0006: bne.un.s IL_000a

        IL_0008: ldc.i4.1
        IL_0009: ret

        IL_000a: ldarg.1
        IL_000b: ldc.i4.3
        IL_000c: beq.s IL_0012

        IL_000e: ldarg.1
        IL_000f: ldc.i4.4
        IL_0010: bne.un.s IL_0014

        IL_0012: ldc.i4.2
        IL_0013: ret

        IL_0014: ldc.i4.3
        IL_0015: ret
```