@dsyme 2021-06-01
I did some further work and the code for `if (x = 1 || x = 2) then ...` is now identical to C#.

As expected a whole bunch of baselines need updating, making a list of the first round of them here for reference

```
2021-06-01T18:18:36.0805760Z CodeGen\EmittedIL\GeneratedIterators (GenIter01.fs -) -- failed
2021-06-01T18:18:36.0815887Z CodeGen\EmittedIL\GeneratedIterators (GenIter02.fs -) -- failed
2021-06-01T18:18:36.0818775Z CodeGen\EmittedIL\GeneratedIterators (GenIter03.fs -) -- failed
2021-06-01T18:18:36.0823673Z CodeGen\EmittedIL\GeneratedIterators (GenIter04.fs -) -- failed
2021-06-01T18:18:36.1054804Z CodeGen\EmittedIL\InequalityComparison (if (x > y) then ... else ...) -- failed
2021-06-01T18:18:41.6798713Z CodeGen\EmittedIL\ListExpressionStepping (ListExpressionSteppingTest1.fs -) -- failed
2021-06-01T18:18:41.6800379Z CodeGen\EmittedIL\ListExpressionStepping (ListExpressionSteppingTest2.fs -) -- failed
2021-06-01T18:18:41.6801314Z CodeGen\EmittedIL\ListExpressionStepping (ListExpressionSteppingTest3.fs -) -- failed
2021-06-01T18:18:41.6802123Z CodeGen\EmittedIL\ListExpressionStepping (ListExpressionSteppingTest4.fs -) -- failed
2021-06-01T18:18:41.6802935Z CodeGen\EmittedIL\ListExpressionStepping (ListExpressionSteppingTest5.fs -) -- failed
2021-06-01T18:18:41.6803774Z CodeGen\EmittedIL\ListExpressionStepping (ListExpressionSteppingTest6.fs -) -- failed
2021-06-01T18:18:29.3494705Z CodeGen\EmittedIL\CCtorDUWithMember (CCtorDUWithMember01.fs -) -- failed
2021-06-01T18:18:56.9232810Z CodeGen\EmittedIL\Misc (AnonRecd.fs) -- failed
2021-06-01T18:18:56.9247055Z CodeGen\EmittedIL\Misc (CodeGenRenamings01.fs -) -- failed
2021-06-01T18:18:56.9343146Z CodeGen\EmittedIL\Misc (EntryPoint01.fs) -- failed
2021-06-01T18:18:56.9359901Z CodeGen\EmittedIL\Misc (EqualsOnUnions01.fs -) -- failed
2021-06-01T18:18:56.9532526Z CodeGen\EmittedIL\Misc (IfThenElse01.fs) -- failed
2021-06-01T18:18:56.9560892Z CodeGen\EmittedIL\Misc (LetIfThenElse01.fs -) -- failed
2021-06-01T18:18:56.9580032Z CodeGen\EmittedIL\Misc (Lock01.fs -) -- failed
2021-06-01T18:18:56.9654414Z CodeGen\EmittedIL\Misc (Seq_for_all01.fs) -- failed
2021-06-01T18:18:56.9673569Z CodeGen\EmittedIL\Misc (StructsAsArrayElements01.fs -) -- failed
2021-06-01T18:18:56.9728429Z CodeGen\EmittedIL\Misc (TryWith_NoFilterBlocks01.fs) -- failed
2021-06-01T18:18:56.9783557Z CodeGen\EmittedIL\Misc (Structs01.fs -) -- failed
2021-06-01T18:18:56.9817744Z CodeGen\EmittedIL\Misc (Structs02.fs -) -- failed
2021-06-01T18:18:57.0006555Z CodeGen\EmittedIL\Misc (GeneralizationOnUnions01.fs) -- failed
2021-06-01T18:18:57.1932730Z CodeGen\EmittedIL\QueryExpressionStepping (Linq101Aggregates01.fs - CodeGen) -- failed
2021-06-01T18:18:57.1987955Z CodeGen\EmittedIL\QueryExpressionStepping (Linq101ElementOperators01.fs - CodeGen) -- failed
2021-06-01T18:18:57.2080815Z CodeGen\EmittedIL\QueryExpressionStepping (Linq101Joins01.fs - CodeGen) -- failed
2021-06-01T18:18:57.2107180Z CodeGen\EmittedIL\QueryExpressionStepping (Linq101Ordering01.fs - CodeGen) -- failed
2021-06-01T18:18:57.2136342Z CodeGen\EmittedIL\QueryExpressionStepping (Linq101Partitioning01.fs - CodeGen) -- failed
2021-06-01T18:18:57.2192864Z CodeGen\EmittedIL\QueryExpressionStepping (Linq101Quantifiers01.fs - CodeGen) -- failed
2021-06-01T18:18:57.2229203Z CodeGen\EmittedIL\QueryExpressionStepping (Linq101Select01.fs - CodeGen) -- failed
2021-06-01T18:18:57.2376926Z CodeGen\EmittedIL\QueryExpressionStepping (Linq101SetOperators01.fs - CodeGen) -- failed
2021-06-01T18:18:57.2529621Z CodeGen\EmittedIL\QueryExpressionStepping (Linq101Where01.fs - CodeGen) -- failed
2021-06-01T18:18:57.3027540Z CodeGen\EmittedIL\SeqExpressionStepping (SeqExpressionSteppingTest1.fs -) -- failed
2021-06-01T18:18:57.3043637Z CodeGen\EmittedIL\SeqExpressionStepping (SeqExpressionSteppingTest2.fs -) -- failed
2021-06-01T18:18:57.3059261Z CodeGen\EmittedIL\SeqExpressionStepping (SeqExpressionSteppingTest3.fs -) -- failed
2021-06-01T18:18:57.3091176Z CodeGen\EmittedIL\SeqExpressionStepping (SeqExpressionSteppingTest4.fs -) -- failed
2021-06-01T18:18:57.3115392Z CodeGen\EmittedIL\SeqExpressionStepping (SeqExpressionSteppingTest5.fs -) -- failed
2021-06-01T18:18:57.3161785Z CodeGen\EmittedIL\SeqExpressionStepping (SeqExpressionSteppingTest6.fs -) -- failed
2021-06-01T18:18:57.3189831Z CodeGen\EmittedIL\SeqExpressionStepping (SeqExpressionSteppingTest7.fs -) -- failed
2021-06-01T18:18:57.3718141Z CodeGen\EmittedIL\SeqExpressionTailCalls (SeqExpressionTailCalls01.fs -) -- failed
2021-06-01T18:18:57.3758408Z CodeGen\EmittedIL\SeqExpressionTailCalls (SeqExpressionTailCalls02.fs -) -- failed
2021-06-01T18:18:57.4704783Z CodeGen\EmittedIL\StaticInit (StaticInit_Struct01.fs -) -- failed
2021-06-01T18:18:57.4725286Z CodeGen\EmittedIL\StaticInit (StaticInit_Class01.fs -) -- failed
2021-06-01T18:18:57.5129646Z CodeGen\EmittedIL\SteppingMatch (SteppingMatch03.fs) -- failed
2021-06-01T18:18:57.5393494Z CodeGen\EmittedIL\SteppingMatch (SteppingMatch04.fs) -- failed
2021-06-01T18:18:57.5500292Z CodeGen\EmittedIL\SteppingMatch (SteppingMatch05.fs) -- failed
2021-06-01T18:18:57.5690359Z CodeGen\EmittedIL\SteppingMatch (SteppingMatch06.fs -) -- failed
2021-06-01T18:18:57.5713454Z CodeGen\EmittedIL\SteppingMatch (SteppingMatch07.fs -) -- failed
2021-06-01T18:18:57.5810766Z CodeGen\EmittedIL\SteppingMatch (SteppingMatch09.fs) -- failed
2021-06-01T18:19:18.3021365Z CodeGen\EmittedIL\TestFunctions (TestFunction16.fs -) -- failed
2021-06-01T18:19:18.3069715Z CodeGen\EmittedIL\TestFunctions (TestFunction17.fs -) -- failed
2021-06-01T18:19:18.3106746Z CodeGen\EmittedIL\TestFunctions (TestFunction21.fs -) -- failed
2021-06-01T18:19:18.3190270Z CodeGen\EmittedIL\TestFunctions (TestFunction3b.fs -) -- failed
2021-06-01T18:19:18.3228101Z CodeGen\EmittedIL\TestFunctions (TestFunction3c.fs -) -- failed
2021-06-01T18:19:18.3291150Z CodeGen\EmittedIL\TestFunctions (TestFunction8.fs) -- failed
2021-06-01T18:19:18.3355678Z CodeGen\EmittedIL\TestFunctions (TestFunction9.fs) -- failed
2021-06-01T18:19:18.3374826Z CodeGen\EmittedIL\TestFunctions (TestFunction9b.fs) -- failed
2021-06-01T18:19:18.3378883Z CodeGen\EmittedIL\TestFunctions (TestFunction9b1.fs) -- failed
2021-06-01T18:19:18.3472270Z CodeGen\EmittedIL\TestFunctions (TestFunction9b2.fs) -- failed
2021-06-01T18:19:18.3480985Z CodeGen\EmittedIL\TestFunctions (TestFunction9b3.fs) -- failed
2021-06-01T18:19:18.3613618Z CodeGen\EmittedIL\TestFunctions (TestFunction22f.fs) -- failed
2021-06-01T18:19:18.3615767Z CodeGen\EmittedIL\TestFunctions (TestFunction22g.fs) -- failed
2021-06-01T18:19:18.3627411Z CodeGen\EmittedIL\TestFunctions (TestFunction24.fs -) -- failed
```