@dominikprzywara 2021-06-02
NEW
```
BenchmarkDotNet=v0.12.1, OS=Windows 10.0.19042
AMD Ryzen 9 5900X, 1 CPU, 24 logical and 12 physical cores
.NET Core SDK=5.0.104
  [Host]     : .NET Core 5.0.6 (CoreCLR 5.0.621.22011, CoreFX 5.0.621.22011), X64 RyuJIT DEBUG
  DefaultJob : .NET Core 5.0.6 (CoreCLR 5.0.621.22011, CoreFX 5.0.621.22011), X64 RyuJIT


| Method | x |      Mean |     Error |    StdDev | Code Size | BranchInstructions/Op | BranchMispredictions/Op |
|------- |-- |----------:|----------:|----------:|----------:|----------------------:|------------------------:|
| CSharp | 1 | 0.6229 ns | 0.0033 ns | 0.0028 ns |      45 B |                     2 |                       0 |
| FSharp | 1 | 0.6217 ns | 0.0032 ns | 0.0030 ns |      45 B |                     2 |                       0 |
| CSharp | 2 | 0.6732 ns | 0.0030 ns | 0.0028 ns |      45 B |                     2 |                       0 |
| FSharp | 2 | 0.6759 ns | 0.0037 ns | 0.0033 ns |      45 B |                     2 |                       0 |
| CSharp | 3 | 0.8321 ns | 0.0046 ns | 0.0043 ns |      45 B |                     2 |                       0 |
| FSharp | 3 | 1.0849 ns | 0.0043 ns | 0.0038 ns |      45 B |                     2 |                       0 |
| CSharp | 4 | 0.8783 ns | 0.0057 ns | 0.0053 ns |      45 B |                     2 |                       0 |
| FSharp | 4 | 0.8782 ns | 0.0034 ns | 0.0030 ns |      45 B |                     2 |                       0 |
```

OLD
```
BenchmarkDotNet=v0.12.1, OS=Windows 10.0.19042
AMD Ryzen 9 5900X, 1 CPU, 24 logical and 12 physical cores
.NET Core SDK=5.0.104
  [Host]     : .NET Core 5.0.6 (CoreCLR 5.0.621.22011, CoreFX 5.0.621.22011), X64 RyuJIT DEBUG
  DefaultJob : .NET Core 5.0.6 (CoreCLR 5.0.621.22011, CoreFX 5.0.621.22011), X64 RyuJIT


| Method | x |      Mean |     Error |    StdDev | Code Size | BranchInstructions/Op | BranchMispredictions/Op |
|------- |-- |----------:|----------:|----------:|----------:|----------------------:|------------------------:|
| CSharp | 1 | 0.6199 ns | 0.0049 ns | 0.0043 ns |      45 B |                     2 |                       0 |
| FSharp | 1 | 0.7116 ns | 0.0033 ns | 0.0029 ns |      45 B |                     2 |                       0 |
| CSharp | 2 | 0.6744 ns | 0.0055 ns | 0.0051 ns |      45 B |                     2 |                       0 |
| FSharp | 2 | 1.0823 ns | 0.0051 ns | 0.0047 ns |      45 B |                     2 |                       0 |
| CSharp | 3 | 1.0834 ns | 0.0034 ns | 0.0032 ns |      45 B |                     2 |                       0 |
| FSharp | 3 | 0.9333 ns | 0.0292 ns | 0.0274 ns |      45 B |                     2 |                       0 |
| CSharp | 4 | 0.6257 ns | 0.0025 ns | 0.0022 ns |      45 B |                     2 |                       0 |
| FSharp | 4 | 1.2912 ns | 0.0043 ns | 0.0041 ns |      45 B |                     3 |                       0 |
```