@kerams 2021-06-02
@dsyme, @badamczewski, here you go, a 2019 Ryzen CPU.

Old:

``` ini

BenchmarkDotNet=v0.12.1, OS=Windows 10.0.19043
AMD Ryzen 7 3700X, 1 CPU, 16 logical and 8 physical cores
.NET Core SDK=5.0.300-preview.21258.4
  [Host]     : .NET Core 5.0.5 (CoreCLR 5.0.521.16609, CoreFX 5.0.521.16609), X64 RyuJIT DEBUG
  DefaultJob : .NET Core 5.0.5 (CoreCLR 5.0.521.16609, CoreFX 5.0.521.16609), X64 RyuJIT


```
| Method | x |      Mean |     Error |    StdDev | Code Size | BranchInstructions/Op | BranchMispredictions/Op |
|------- |-- |----------:|----------:|----------:|----------:|----------------------:|------------------------:|
| **CSharp** | **1** | **0.6838 ns** | **0.0091 ns** | **0.0085 ns** |      **45 B** |                     **2** |                       **0** |
| FSharp | 1 | 0.4287 ns | 0.0011 ns | 0.0009 ns |      45 B |                     2 |                       0 |
| **CSharp** | **2** | **0.4741 ns** | **0.0021 ns** | **0.0017 ns** |      **45 B** |                     **2** |                       **0** |
| FSharp | 2 | 0.8899 ns | 0.0055 ns | 0.0049 ns |      45 B |                     2 |                       0 |
| **CSharp** | **3** | **0.8931 ns** | **0.0022 ns** | **0.0019 ns** |      **45 B** |                     **3** |                       **0** |
| FSharp | 3 | 0.6774 ns | 0.0057 ns | 0.0053 ns |      45 B |                     2 |                       0 |
| **CSharp** | **4** | **0.7042 ns** | **0.0146 ns** | **0.0114 ns** |      **45 B** |                     **2** |                       **0** |
| FSharp | 4 | 1.3445 ns | 0.0008 ns | 0.0007 ns |      45 B |                     2 |                       0 |

New:

``` ini

BenchmarkDotNet=v0.12.1, OS=Windows 10.0.19043
AMD Ryzen 7 3700X, 1 CPU, 16 logical and 8 physical cores
.NET Core SDK=5.0.300-preview.21258.4
  [Host]     : .NET Core 5.0.5 (CoreCLR 5.0.521.16609, CoreFX 5.0.521.16609), X64 RyuJIT DEBUG
  DefaultJob : .NET Core 5.0.5 (CoreCLR 5.0.521.16609, CoreFX 5.0.521.16609), X64 RyuJIT


```
| Method | x |      Mean |     Error |    StdDev | Code Size | BranchInstructions/Op | BranchMispredictions/Op |
|------- |-- |----------:|----------:|----------:|----------:|----------------------:|------------------------:|
| **CSharp** | **1** | **1.1253 ns** | **0.0009 ns** | **0.0008 ns** |      **45 B** |                     **2** |                       **0** |
| FSharp | 1 | 1.1480 ns | 0.0021 ns | 0.0016 ns |      45 B |                     2 |                       0 |
| **CSharp** | **2** | **0.4583 ns** | **0.0007 ns** | **0.0006 ns** |      **45 B** |                     **2** |                       **0** |
| FSharp | 2 | 0.4763 ns | 0.0053 ns | 0.0050 ns |      45 B |                     2 |                       0 |
| **CSharp** | **3** | **0.9002 ns** | **0.0021 ns** | **0.0016 ns** |      **45 B** |                     **2** |                       **0** |
| FSharp | 3 | 0.9221 ns | 0.0021 ns | 0.0020 ns |      45 B |                     2 |                       0 |
| **CSharp** | **4** | **1.1160 ns** | **0.0009 ns** | **0.0008 ns** |      **45 B** |                     **2** |                       **0** |
| FSharp | 4 | 0.6995 ns | 0.0039 ns | 0.0033 ns |      45 B |                     2 |                       0 |

If anyone wants to run this with a preview of VS, the correct tool path is `C:\Program Files (x86)\Microsoft Visual Studio\2019\Preview\Common7\IDE\CommonExtensions\Microsoft\FSharp\Tools`. Also had to disable virtualization in the BIOS, otherwise the benchmark would not run.